+++ src/command_modules/azure-cli-network/HISTORY.rst
@@ -3,6 +3,8 @@
 Release History
 ===============
 
+* `application-gateway`: Add `rewrite-rule` command group.
+
 2.3.4
 +++++
 * `vpn-connection update`: Fix issue where updating a VPN connection between gateways in different subscriptions would fail.